backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
pydantic>=2.6.0
python-multipart>=0.0.6
openpyxl>=3.1.2
google-api-python-client>=2.116.0
google-auth>=2.27.0
PyJWT>=2.8.0
slowapi>=0.1.9
Pillow>=10.2.0
